% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumption_check.R
\name{assumption_check}
\alias{assumption_check}
\title{Create tables and graphs providing cell means model assumption checking}
\usage{
assumption_check(x, y, data, digits = 3, table = c(TRUE, FALSE),
  graph = c(TRUE, FALSE), ...)
}
\arguments{
\item{x}{a vector that contains different levels whose variances are to be compared}

\item{y}{a numeric vector of data values}

\item{data}{a data frame, list or environment (or object coercible by as.data.frame to a data frame)}

\item{digits}{the digits of the data values (default is 3)}

\item{table}{a boolean value specifying providing table or not, must be TRUE (default) or FALSE}

\item{graph}{a boolean value specifying providing graph or not, must be TRUE (default) or FALSE}

\item{...}{further arguments to be passed.}
}
\value{
By default, it returns both tables and graphs. Tables contain results of Shapiro-Wilk normality test and Bartlett test. Graphs are qq-plot, boxplot, residual plot and histogram. If only asking for graphs, specify table = False. If only asking for tables, specify graph = False.
}
\description{
Based on cell means model, model assumption checking was performed by displaying tables containing test results and graphs. Normality and heteroscedasticity are examined in this function. Shapiro-Wilk normality test, qq-plot and histogram are provided to test normality. Bartlett test, residual plot and boxplot are for heteroscedasticity test.
}
\examples{
x <- c (1,1,1,1,2,2,2,2,3,3,3,3)

y <- c (1,3,3,2,4,3,2,3,2,3,3,2)

data <- data.frame(cbind(x,y))

assumption_check(x,y,data)

}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.
}
\seealso{
\code{\link[stats]{shapiro.test}}

\code{\link[stats]{bartlett.test}}
}
