% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sum.R
\name{data_sum}
\alias{data_sum}
\title{Provide basic summary statistics}
\usage{
data_sum(y)
}
\arguments{
\item{y}{An object for which a method has been defined, or a numeric vector containing the values whose length, mean, median, quartiles are to be computed.}
}
\value{
The method returns \code{y} containing the length, mean, median, 25% and 75% quartiles of the object. Missing values are removed by default.
}
\description{
Generic function to provide basic summary statistics analysis for numeric vectors
}
\examples{
x<-c(0:10,50)
data_sum(x)
}
\references{
Becker, R.A., Chambers, J.M., and Wilks, A.R.(1988) The New S Language. Wadsworth & Brooks/Cole.
}
\seealso{
\code{\link[base]{length}}, \code{\link[base]{mean}},\code{\link[stats]{quantile}}
}
