% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gln.R
\docType{package}
\name{gln}
\alias{gln}
\alias{gln-package}
\title{gln: A package for analysis of variance(ANOVA).}
\description{
The "gln" is a set of functions that analysis whether there is difference in any of the means among different groups in our dataset. This package is designed to make it easy to install and load multiple "ANOVA" functions in a single step. The “gln” packages includes parametric (F-Test) and non-parametric method (KW-Test). The output contains both tables and figures based on the result of ANOVA analysis.
}
\details{
Package: gln

Version:0.1.0

Date:2018-3-1

Depends:R(>=3.2.0)

Imports: multcomp, tidyverse

Suggests:formatR, knitr, rmarkdown
}
\section{Functions}{

The gln package provides five categories of important functions:

data_sum: \code{\link[gln]{data_sum}}

all_sum_test: \code{\link[gln]{all_sum_test}}
}

\section{gln functions}{

}

\author{
Meng Lan   menglan36@gmail.com,

        Xinye Gui  leafly80@163.com,

        Yuwei Ni   yuweiniCornell@gmail.com
}
