% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Provide basic summary statistics
Generic function to provide basic summary statistics for numeric vectors}
\usage{
hello(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Provide basic summary statistics
Generic function to provide basic summary statistics for numeric vectors
}
\examples{
hello(1, 1)
hello(10, 1)

}
