% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gln.R
\docType{package}
\name{gln}
\alias{gln}
\alias{gln-package}
\title{gln: Numerical and Graphical Summary for Comparing Continuous Data in Multiple Groups}
\description{
This package gives a gallery of well-formatted summary tables and graphs which can be made using the gln package for comparing continuous data in multiple groups. To check the model assumptionm, tables containing test results and graphs are displayed for normality and heteroscedasticity test.To examine whether there is difference across the groups, overall tables and graphs are provided for comparison using parametric and/or nonparametric approach. To detect between which groups the difference is, the package produces different tables for post hoc analysis depending on one factor or two factors in the data.
}
\details{
Package: gln

Version:0.1.0

Date:2018-3-1

Depends:R(>=3.2.0)

Imports: multcomp, tidyverse,devtools

Suggests:formatR, knitr, rmarkdown
}
\section{Functions}{

The gln package provides three categories of important functions:


all_sum_test: \code{\link[gln]{all_sum_test}}

posthoc_test: \code{\link[gln]{posthoc_test}}

assumption_check:\code{\link[gln]{assumption_check}}
}

\section{Data}{

The gln package provide one data set to analysis or test the package, which names Mousedata. It is an analyze data used to evaluate leves of urinary PGE-M and and PGD-M in obese vs. lean mice +/- celecoxib.

Mousedata: \code{\link[gln]{Mousedata}}
}

\section{gln functions}{

}

\author{
Meng Lan   menglan36@gmail.com,

        Xinye Gui  leafly80@163.com,

        Yuwei Ni   yuweiniCornell@gmail.com
}
